requirements.txt
crewai>=0.41.0
crewai-tools>=0.8.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
jupyter>=1.0.0
ipython>=8.12.0
colorama>=0.4.6
rich>=13.7.0
